Move fileclient tests to tests/integration/fileserver/fileclient_test.py
Being in the root of tests/integration/, these were not being run in
jenkins and would only run if executed manually using the -n flag in
runtests.py.
